FuncionalidadesNuevas 20180702

+++ app/src/main/java/com/example/a201495_2/porkgestion/bo_clases/SanidadCerdo.java
@@ -12,14 +12,16 @@ public class SanidadCerdo {
     private Context appContext;
     private dataBaseOpenHelper dbAcces;
     private int idCerdo, idSanidad;
-    private double dDosis;
+    private int idCerdoPK, idSanidadPK;
+    private String strDosis;
+    private String strFechaAdministracionPK;
     private String strTipoMedicamento, strNombreMedicamento, strViaAdministracion, strFechaAdministracion;
     private String strCodigoCerdo, strSexoCerdo, strError;
 
-    public SanidadCerdo(int idCerdo, int idSanidad, double dDosis, String strTipoMedicamento, String strNombreMedicamento, String strViaAdministracion, String strFechaAdministracion, String strCodigoCerdo, String strSexoCerdo) {
+    public SanidadCerdo(int idCerdo, int idSanidad, String strDosis, String strTipoMedicamento, String strNombreMedicamento, String strViaAdministracion, String strFechaAdministracion, String strCodigoCerdo, String strSexoCerdo) {
         this.idCerdo = idCerdo;
         this.idSanidad = idSanidad;
-        this.dDosis = dDosis;
+        this.strDosis = strDosis;
         this.strTipoMedicamento = strTipoMedicamento;
         this.strNombreMedicamento = strNombreMedicamento;
         this.strViaAdministracion = strViaAdministracion;
@@ -27,6 +29,8 @@ public class SanidadCerdo {
         this.strCodigoCerdo = strCodigoCerdo;
         this.strSexoCerdo = strSexoCerdo;
     }
+
+
 
     public SanidadCerdo(Context appContext) {
         this.appContext = appContext;
@@ -48,12 +52,12 @@ public class SanidadCerdo {
         this.idSanidad = idSanidad;
     }
 
-    public double getdDosis() {
-        return dDosis;
+    public String getStrDosis() {
+        return strDosis;
     }
 
-    public void setdDosis(double dDosis) {
-        this.dDosis = dDosis;
+    public void setStrDosis(String strDosis) {
+        this.strDosis = strDosis;
     }
 
     public String getStrTipoMedicamento() {
@@ -108,6 +112,18 @@ public class SanidadCerdo {
         return strError;
     }
 
+    public void setIdCerdoPK(int idCerdoPK) {
+        this.idCerdoPK = idCerdoPK;
+    }
+
+    public void setIdSanidadPK(int idSanidadPK) {
+        this.idSanidadPK = idSanidadPK;
+    }
+
+    public void setStrFechaAdministracionPK(String strFechaAdministracionPK) {
+        this.strFechaAdministracionPK = strFechaAdministracionPK;
+    }
+
     public boolean insertSanidadCerdo(){
         dbAcces = new dataBaseOpenHelper(appContext);
         ContentValues ctValores = new ContentValues();
@@ -115,6 +131,7 @@ public class SanidadCerdo {
         ctValores.put("IDSANIDAD",this.idSanidad);
         ctValores.put("VIAADMINISTRACION",this.strViaAdministracion);
         ctValores.put("FECHAADMINISTRACION",this.strFechaAdministracion);
+        ctValores.put("DOSIS",this.strDosis);
         dbAcces.insertDatabase("SANIDADCERDO",ctValores);
         this.strError = dbAcces.getErrorDB();
         return dbAcces.getErrorDB()==null;
@@ -127,28 +144,27 @@ public class SanidadCerdo {
         ctValores.put("IDSANIDAD",this.idSanidad);
         ctValores.put("VIAADMINISTRACION",this.strViaAdministracion);
         ctValores.put("FECHAADMINISTRACION",this.strFechaAdministracion);
-        String strArgs[] = new String[]{String.valueOf(this.idSanidad),String.valueOf(this.idCerdo)};
-        dbAcces.updateDatabase("SANIDADCERDO",ctValores,"IDSANIDAD=? AND IDCERDO=?",strArgs);
+        ctValores.put("DOSIS",this.strDosis);
+        String strArgs[] = new String[]{String.valueOf(this.idSanidadPK),String.valueOf(this.idCerdoPK),String.valueOf(this.strFechaAdministracionPK)};
+        dbAcces.updateDatabase("SANIDADCERDO",ctValores,"IDSANIDAD=? AND IDCERDO=? AND FECHAADMINISTRACION=?",strArgs);
         this.strError = dbAcces.getErrorDB();
         return dbAcces.getErrorDB()==null;
     }
 
     public Boolean deleteSanidadCerdo(){
         dbAcces = new dataBaseOpenHelper(appContext);
-        String strArgs[] = new String[]{String.valueOf(this.idSanidad),String.valueOf(this.idCerdo)};
-        dbAcces.deleteDatabase ("SANIDADCERDO","IDSANIDAD=? AND IDCERDO=?",strArgs);
+        String strArgs[] = new String[]{String.valueOf(this.idSanidadPK),String.valueOf(this.idCerdoPK),String.valueOf(this.strFechaAdministracionPK)};
+        dbAcces.deleteDatabase ("SANIDADCERDO","IDSANIDAD=? AND IDCERDO=? AND FECHAADMINISTRACION=?",strArgs);
         return dbAcces.getErrorDB()==null;
     }
 
     public SanidadCerdo getSanidadCerdo(int idCerdo){
         dbAcces = new dataBaseOpenHelper(appContext);
-        String strColumns[] = new String[]{"IDSANIDAD","IDCERDO","CODIGO","SEXO","TIPOMEDICAMENTO","NOMBREMEDICAMENTO","FECHAADMINISTRACION","VIAADMINISTRACION"};
-        String strArgs[] = new String[]{String.valueOf(idCerdo)};
-        SanidadCerdo tmpObject = new SanidadCerdo(appContext);
-        String sql="SELECT SANIDAD.IDSANIDAD,CERDO.IDCERDO, CERDO.CODIGO,CERDO.SEXO,SANIDAD.TIPOMEDICAMENTO,SANIDAD.NOMBREMEDICAMENTO,SANIDAD.OBSERVACIONES,SANIDADCERDO.FECHAADMINISTRACION,SANIDADCERDO.VIAADMINISTRACION,SANIDADCERDO.DOSIS FROM SANIDADCERDO INNER JOIN CERDO ON SANIDADCERDO.IDCERDO = CERDO.IDCERDO INNER JOIN SANIDAD ON SANIDAD.IDSANIDAD=SANIDADCERDO.IDSANIDAD";
         Cursor crResult;
+        SanidadCerdo tmpObject = new SanidadCerdo(appContext);
+        String strSql = String.format("SELECT SANIDAD.IDSANIDAD,CERDO.IDCERDO, CERDO.CODIGO,CERDO.SEXO,SANIDAD.TIPOMEDICAMENTO,SANIDAD.NOMBREMEDICAMENTO,SANIDAD.OBSERVACIONES,SANIDADCERDO.FECHAADMINISTRACION,SANIDADCERDO.VIAADMINISTRACION,SANIDADCERDO.DOSIS FROM SANIDADCERDO LEFT JOIN CERDO ON SANIDADCERDO.IDCERDO = CERDO.IDCERDO LEFT JOIN SANIDAD ON SANIDAD.IDSANIDAD=SANIDADCERDO.IDSANIDAD WHERE IDCERDO = '%s'",String.valueOf(idCerdo));
         dbAcces.openDataBase();
-        crResult = dbAcces.qweryDatabaseBySql (sql);
+        crResult = dbAcces.qweryDatabaseBySql (strSql);
         if (crResult.moveToFirst()) {
             tmpObject.setIdSanidad(crResult.getInt(0));
             tmpObject.setIdCerdo(crResult.getInt(1));
@@ -158,7 +174,7 @@ public class SanidadCerdo {
             tmpObject.setStrNombreMedicamento(crResult.getString(5));
             tmpObject.setStrFechaAdministracion(crResult.getString(7));
             tmpObject.setStrViaAdministracion(crResult.getString(8));
-            tmpObject.setdDosis(crResult.getDouble(9));
+            tmpObject.setStrDosis(crResult.getString(9));
         }
         this.strError = dbAcces.getErrorDB();
         dbAcces.closeDataBase();
@@ -167,11 +183,12 @@ public class SanidadCerdo {
 
     public ArrayList<SanidadCerdo> getAllSanidadCerdo(){
         dbAcces = new dataBaseOpenHelper(appContext);
-        String strColumns[] = new String[]{"IDSANIDAD","IDCERDO","CODIGO","SEXO","TIPOMEDICAMENTO","NOMBREMEDICAMENTO","FECHAADMINISTRACION","VIAADMINISTRACION"};
         Cursor crResult;
         ArrayList<SanidadCerdo> listObject = new ArrayList<SanidadCerdo>();
+        String strSsql="SELECT SANIDAD.IDSANIDAD,CERDO.IDCERDO, CERDO.CODIGO,CERDO.SEXO,SANIDAD.TIPOMEDICAMENTO,SANIDAD.NOMBREMEDICAMENTO,SANIDAD.OBSERVACIONES,SANIDADCERDO.FECHAADMINISTRACION,SANIDADCERDO.VIAADMINISTRACION,SANIDADCERDO.DOSIS FROM SANIDADCERDO LEFT JOIN CERDO ON SANIDADCERDO.IDCERDO = CERDO.IDCERDO LEFT JOIN SANIDAD ON SANIDAD.IDSANIDAD=SANIDADCERDO.IDSANIDAD ORDER BY CERDO.IDCERDO";
         dbAcces.openDataBase();
-        crResult = dbAcces.qweryDatabase("VW_SANIDADCERDO", strColumns, null, null, "IDSanidadCerdo");
+        crResult = dbAcces.qweryDatabaseBySql (strSsql);
+
         if (crResult.moveToFirst()) {
             do {
                 SanidadCerdo tmpObject = new SanidadCerdo(appContext);
@@ -181,8 +198,9 @@ public class SanidadCerdo {
                 tmpObject.setStrSexoCerdo(crResult.getString(3));
                 tmpObject.setStrTipoMedicamento(crResult.getString(4));
                 tmpObject.setStrNombreMedicamento(crResult.getString(5));
-                tmpObject.setStrFechaAdministracion(crResult.getString(6));
-                tmpObject.setStrViaAdministracion(crResult.getString(7));
+                tmpObject.setStrFechaAdministracion(crResult.getString(7));
+                tmpObject.setStrViaAdministracion(crResult.getString(8));
+                tmpObject.setStrDosis(crResult.getString(9));
                 listObject.add(tmpObject);
             } while (crResult.moveToNext());
         }
@@ -191,34 +209,6 @@ public class SanidadCerdo {
         return listObject;
     }
 
-    public ArrayList<SanidadCerdo>  getSanidadCerdoByQwery(String strvalor1, String strValor2, String strVaslor3){
-        dbAcces = new dataBaseOpenHelper(appContext);
-        Cursor crResult;
-        String strSql=null;
-        ArrayList<SanidadCerdo> listObject = new ArrayList<SanidadCerdo>();
-        //TODO:  A partir de los los campos recibido, se arma el Qhery y se actualiza la variable strSql;
-        dbAcces.openDataBase();
-        crResult = dbAcces.qweryDatabaseBySql(strSql);
-        //TODO: Crear el arraylist a partir de los resultados
-        if (crResult.moveToFirst()) {
-            do {
-                SanidadCerdo tmpObject = new SanidadCerdo(appContext);
-                tmpObject.setIdSanidad(crResult.getInt(0));
-                tmpObject.setIdCerdo(crResult.getInt(1));
-                tmpObject.setStrCodigoCerdo(crResult.getString(2));
-                tmpObject.setStrSexoCerdo(crResult.getString(3));
-                tmpObject.setStrTipoMedicamento(crResult.getString(4));
-                tmpObject.setStrNombreMedicamento(crResult.getString(5));
-                tmpObject.setStrFechaAdministracion(crResult.getString(6));
-                tmpObject.setStrViaAdministracion(crResult.getString(7));
-                listObject.add(tmpObject);
-            } while (crResult.moveToNext());
-        }
-
-        this.strError = dbAcces.getErrorDB();
-        dbAcces.closeDataBase();
-        return listObject;
-    }
 
     public Boolean execDatabaseByQwery(String strvalor1, String strValor2, String strVaslor3){
         dbAcces = new dataBaseOpenHelper(appContext);
@@ -235,6 +225,20 @@ public class SanidadCerdo {
         Boolean bResult  = false;
         Cursor crResult;
         String strSql = String.format("SELECT COUNT(*) AS TOTAL FROM SanidadCerdo WHERE IDSANIDAD='%s'",idSanidad);
+        dbAcces.openDataBase();
+        crResult = dbAcces.qweryDatabaseBySql(strSql);
+        if (crResult.moveToFirst()) {
+            bResult  =  crResult.getInt(0)>0;
+        }
+        dbAcces.closeDataBase();
+        return bResult;
+    }
+
+    public Boolean existSanidadCerdoByIdSanidadCerdo(int idSanidad, int idCerdo, String strFechaAdmin){
+        dbAcces = new dataBaseOpenHelper(appContext);
+        Boolean bResult  = false;
+        Cursor crResult;
+        String strSql = String.format("SELECT COUNT(*) AS TOTAL FROM SanidadCerdo WHERE IDSANIDAD='%s' AND IDCERDO='%s' AND FECHAADMINISTRACION='%s' ",idSanidad, idCerdo, strFechaAdmin);
         dbAcces.openDataBase();
         crResult = dbAcces.qweryDatabaseBySql(strSql);
         if (crResult.moveToFirst()) {
